Update m_idle.h
git-svn-id: http://miranda.googlecode.com/svn/trunk@401 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/include/m_idle.h
@@ -54,7 +54,7 @@ typedef struct {
 	int cbSize;			     // sizeof()
 	unsigned int idleTime;	 // idle in mins, if zero then disabled
 	int privacy;		     // user doesnt want other people seeing anything more than they are idle
-    unsigned short aaStatus; // status to go to when user is auto away
+    int aaStatus; // status to go to when user is auto away
     
 } MIRANDA_IDLE_INFO;
